+++ src/app/app.component.scss
@@ -6,12 +6,14 @@ mat-sidenav {
   width: 250px;
 }
 
-.sidenav-content-margin{
+.sidenav-content-margin {
   @media screen and (max-width: 599px) {
-    padding-top:56px;
-    height: calc( 100vh - 56px );
+    padding-top: 56px;
+    height: calc(100vh - 56px);
   }
-  padding-top:64px;
-  height: calc( 100vh - 64px );
+  padding-top: 64px;
+  height: calc(100vh - 64px);
   overflow-y: auto;
 }
+
+
