+++ src/app/auth/login/login.component.scss
@@ -1,13 +1,9 @@
-.mat-form-field {
-  min-width: 300px;
-}
-
 .login-form {
-  margin: 10px
-}
+  margin: 10px;
+  .facebook-login {
+    max-width: 500px;
+    background-color: #3b5998;
+    color: white;
+  }
 
-.facebook-login {
-  max-width: 500px;
-  background-color: #3b5998;
-  color: white;
 }
