+++ src/app/auth/login/login.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, NgForm, Validators } from '@angular/forms';
 import { Observable } from 'rxjs/Observable';
 import { Select, Store } from '@ngxs/store';
-import * as authActions from '../auth.state';
+import * as authActions from '../ngxs/auth.classes';
 import { PasswordValidators } from 'ngx-validators';
 
 @Component({
