+++ src/app/auth/ngxs/auth.consts.ts
@@ -1,0 +1,10 @@
+export const START_LOADING = '[Auth] START_LOADING';
+export const STOP_LOADING = '[Auth] STOP_LOADING';
+export const START_EMAIL_SIGNUP = '[Auth] START_EMAIL_SIGNUP';
+export const START_SOCIAL_SIGNUP = '[Auth] START_SOCIAL_SIGNUP';
+export const FINISH_SOCIAL_SIGNUP = '[Auth] FINISH_SOCIAL_SIGNUP';
+export const START_EMAIL_LOGIN = '[Auth] START_EMAIL_LOGIN';
+export const START_SOCIAL_LOGIN = '[Auth] START_SOCIAL_LOGIN';
+export const START_LOGOUT = '[Auth] START_LOGOUT';
+export const SET_USER = '[Auth] SET_USER';
+export const DELETE_USER = '[Auth] DELETE_USER';
