+++ src/app/auth/ngxs/auth.state.ts
@@ -1,77 +1,26 @@
 import { Action, Selector, State, StateContext, Store } from '@ngxs/store';
-import { FinishSignupModel, SignupModel } from './models/signup';
-import { LoginModel } from './models/login';
 import { Router } from '@angular/router';
 import 'rxjs/add/operator/switchMap';
 import { fromPromise } from 'rxjs/observable/fromPromise';
-import { SendFinishUserMetaModel, SendUserMetaModel, StartUserMetaModel } from './models/user-meta';
+import { SendFinishUserMetaModel, SendUserMetaModel, StartUserMetaModel } from '../models/user-meta';
 import { AngularFireAuth } from 'angularfire2/auth';
 import { AngularFirestore } from 'angularfire2/firestore';
 import * as firebase from 'firebase';
+import {
+  DeleteUser,
+  FinishSocialSignup,
+  SetUser, StartEmailLogin,
+  StartEmailSignup,
+  StartLoading, StartLogout, StartSocialLogin,
+  StartSocialSignup,
+  StopLoading
+} from './auth.classes';
 
-export const START_LOADING = '[Auth] START_LOADING';
-export const STOP_LOADING = '[Auth] STOP_LOADING';
-export const START_EMAIL_SIGNUP = '[Auth] START_EMAIL_SIGNUP';
-export const START_SOCIAL_SIGNUP = '[Auth] START_SOCIAL_SIGNUP';
-export const FINISH_SOCIAL_SIGNUP = '[Auth] FINISH_SOCIAL_SIGNUP';
-export const START_EMAIL_LOGIN = '[Auth] START_EMAIL_LOGIN';
-export const START_SOCIAL_LOGIN = '[Auth] START_SOCIAL_LOGIN';
-export const START_LOGOUT = '[Auth] START_LOGOUT';
-export const SET_USER = '[Auth] SET_USER';
-export const DELETE_USER = '[Auth] DELETE_USER';
 
 export interface AuthStateModel {
   isLoading: boolean;
   isAuthenticated: boolean;
   user: firebase.User | null;
-}
-
-export class StartLoading {
-  static readonly type = START_LOADING;
-}
-
-export class StopLoading {
-  static readonly type = STOP_LOADING;
-}
-
-export class StartEmailSignup {
-  static readonly type = START_EMAIL_SIGNUP;
-
-  constructor(public payload: SignupModel) { }
-}
-
-export class StartSocialSignup {
-  static readonly type = START_SOCIAL_SIGNUP;
-}
-
-export class FinishSocialSignup {
-  static readonly type = FINISH_SOCIAL_SIGNUP;
-
-  constructor(public payload: FinishSignupModel) { }
-}
-
-export class StartEmailLogin {
-  static readonly type = START_EMAIL_LOGIN;
-
-  constructor(public payload: LoginModel) { }
-}
-
-export class StartSocialLogin {
-  static readonly type = START_SOCIAL_LOGIN;
-}
-
-export class StartLogout {
-  static readonly type = START_LOGOUT;
-}
-
-export class SetUser {
-  static readonly type = SET_USER;
-
-  constructor(public payload: firebase.User | null) { }
-}
-
-export class DeleteUser {
-  static readonly type = DELETE_USER;
 }
 
 @State<AuthStateModel>({
@@ -87,9 +36,7 @@ export class AuthState {
   constructor(private router: Router,
               private store: Store,
               private afAuth: AngularFireAuth,
-              private db: AngularFirestore) {
-
-  }
+              private db: AngularFirestore) { }
 
   @Selector()
   static isLoading(state: AuthStateModel) {
@@ -249,6 +196,7 @@ export class AuthState {
   }
 
   private signIn(user: firebase.User | null) {
+    debugger;
     this.store.dispatch(new SetUser(user));
     this.router.navigate(['test']);
   }
