+++ src/app/auth/no-auth.guard.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { Select, Store } from '@ngxs/store';
 import 'rxjs/add/operator/do';
-import { AuthState } from './auth.state';
+import { AuthState } from './ngxs/auth.state';
 import 'rxjs/add/operator/map';
 import { map, take, tap } from 'rxjs/operators';
 import { AngularFireAuth } from 'angularfire2/auth';
@@ -30,6 +30,7 @@ export class NoAuthGuard implements CanActivate, CanLoad {
       take(1),
       map((authState) => !authState),
       tap(notAuthenticated => {
+        debugger;
         if (!notAuthenticated) {
           this.router.navigate(['/test']);
         }
