+++ src/app/auth/signup/finish-signup/finish-signup.component.scss
@@ -1,16 +1,11 @@
-.mat-form-field {
-  min-width: 240px;
-}
-
-.app-place-search {
-  width: 100%;
-  max-width: 600px;
-}
-
 .signup-form {
-  margin: 10px
-}
+  margin: 10px;
+  .app-place-search {
+    width: 100%;
+    max-width: 600px;
+  }
+  .gps-form-field {
+    min-width: 400px;
+  }
 
-.gps-form-field {
-  min-width: 400px;
 }
