+++ src/app/auth/signup/finish-signup/finish-signup.component.ts
@@ -1,7 +1,7 @@
 import { Component, ElementRef, EventEmitter, OnInit, ViewChild } from '@angular/core';
 import { FormControl, FormGroup, NgForm, Validators } from '@angular/forms';
 import { Select, Store } from '@ngxs/store';
-import * as authActions from '../../auth.state';
+import * as authActions from '../../ngxs/auth.classes';
 import { Observable } from 'rxjs/Observable';
 import { LatLngLiteral } from '@agm/core';
 import { PasswordValidators } from 'ngx-validators';
