+++ src/app/auth/signup/signup.component.ts
@@ -2,7 +2,7 @@ import { Component, ElementRef, EventEmitter, OnInit, ViewChild } from '@angular
 import { Observable } from 'rxjs/Observable';
 import { FormControl, FormGroup, NgForm, Validators } from '@angular/forms';
 import { Select, Store } from '@ngxs/store';
-import * as authActions from '../auth.state';
+import * as authActions from '../ngxs/auth.classes';
 import { LatLngLiteral } from '@agm/core';
 import { PasswordValidators } from 'ngx-validators';
