+++ src/app/common/card-post/card-post.component.scss
@@ -1,7 +1,16 @@
-.example-card {
-  width: 20%;
-}
-
 mat-card {
+  width: 20%;
   min-width: 400px;
+  .arrow-upward {
+    color: green;
+  }
+
+  .arrow-downward {
+    color: red;
+  }
+
+  .card-description {
+    width: 90%;
+    text-align: center;
+  }
 }
