+++ src/app/common/chips-select/chips-select.component.html
@@ -1,4 +1,4 @@
-<mat-form-field class="example-chip-list">
+<mat-form-field class="chip-list">
   <mat-chip-list #chipList>
     <mat-chip
       *ngFor="let chip of chips"
