+++ src/app/common/chips-select/chips-select.component.scss
@@ -1,3 +1,3 @@
-.example-chip-list {
+.chip-list {
   width: 100%;
 }
