+++ src/app/common/feed/feed.component.scss
@@ -2,7 +2,3 @@
   height: 100%;
   overflow: scroll;
 }
-
-.filter-select {
-  max-width: 200px;
-}
