+++ src/app/navigation/header/header.component.scss
@@ -1,3 +1,15 @@
+.logo {
+  width: 30px
+}
+
+.join-us {
+  background-color: #3fb5a3
+}
+
+.add-post {
+  background-color: #3fb5a3
+}
+
 a {
   text-decoration: none;
   color: white;
@@ -18,5 +30,5 @@ a:hover, a:active {
   top: 0;
   left: 0;
   z-index: 2;
-  width: 100%  !important;
+  width: 100% !important;
 }
