+++ src/app/welcome/welcome.component.html
@@ -33,7 +33,7 @@
     <!--<h1>Mire vársz még?</h1>-->
     <h1>What are you waiting for?</h1>
     <p>
-      <button mat-raised-button style="background-color:#3fb5a3;color:white" routerLink="/signup">
+      <button mat-raised-button class="join-us" routerLink="/signup">
         Join us
       </button>
     </p>
