+++ src/app/welcome/welcome.component.scss
@@ -1,4 +1,9 @@
-.welcome{
+.welcome {
   text-align: center;
-  margin:0 10px;
+  margin: 0 10px;
+}
+
+.join-us {
+  background-color: #3fb5a3;
+  color: white
 }
